datascripts/requirements.txt
# AI 도서 추천 시스템 - Python 의존성
langchain==0.1.0
langchain-openai==0.0.5
langchain-chroma==0.1.0
openai==1.12.0
chromadb==0.4.22
mysql-connector-python==8.3.0
python-dotenv==1.0.0
numpy==1.26.3
pandas==2.2.0
tiktoken==0.5.2